requirements.txt
click==8.1.7
mysql-connector-python==8.3.0
tabulate==0.9.0
